Rename tests igraph module to networkx
