+++ tests/graph/networkx/__init__.py
@@ -1,5 +1,5 @@
 """
-Functionality from igraph to produce pixelator Graph
+Functionality from networkx to produce pixelator Graph
 of certain types required for testing
 
 Copyright (c) 2023 Pixelgen Technologies AB.
